sensor: max17055: change register access function addr to uint8_t
The register space for the device is on 8 bit, use uint8_t for register
address to the low level read and write functions.

+++ drivers/sensor/max17055/max17055.c
@@ -27,7 +27,7 @@ LOG_MODULE_REGISTER(max17055, CONFIG_SENSOR_LOG_LEVEL);
  * @param valp Place to put the value on success
  * @return 0 if successful, or negative error code from I2C API
  */
-static int max17055_reg_read(const struct device *dev, int reg_addr,
+static int max17055_reg_read(const struct device *dev, uint8_t reg_addr,
 			     int16_t *valp)
 {
 	const struct max17055_config *config = dev->config;
@@ -44,13 +44,13 @@ static int max17055_reg_read(const struct device *dev, int reg_addr,
 	return 0;
 }
 
-static int max17055_reg_write(const struct device *dev, int reg_addr,
+static int max17055_reg_write(const struct device *dev, uint8_t reg_addr,
 			      uint16_t val)
 {
 	const struct max17055_config *config = dev->config;
 	uint8_t buf[3];
 
-	buf[0] = (uint8_t)reg_addr;
+	buf[0] = reg_addr;
 	sys_put_le16(val, &buf[1]);
 
 	return i2c_write_dt(&config->i2c, buf, sizeof(buf));
